fix: constants.ts restructuring
