+++ packages/api/src/auth/decorators/cookiePayload.decorator.ts
@@ -1,15 +1,15 @@
 import { createParamDecorator, ExecutionContext, UnauthorizedException } from '@nestjs/common';
 import jsonwebtoken from 'jsonwebtoken';
-import { TokenConfig } from '../../utils/constants';
 import { Payload } from '../../interfaces/token-types';
 import { CookieRequest } from '../../interfaces/cookie-types';
-import { getCookies } from '../../utils/helpers';
+import { getCookies, constants } from '../../utils';
 
 export const CookiePayload = createParamDecorator(
   (data: unknown, context: ExecutionContext): Payload => {
+    const { tokenConfig } = constants;
     const request = context.switchToHttp().getRequest<CookieRequest>();
     const cookies = getCookies(request);
-    const refresh_token = cookies[TokenConfig.refreshToken.name];
+    const refresh_token = cookies[tokenConfig.refreshToken.name];
     const decode = jsonwebtoken.decode(refresh_token) as Payload;
     if (!decode) {
       throw new UnauthorizedException('Invalid token payload');
