+++ packages/api/src/auth/guards/only-authenticated.guard.ts
@@ -1,21 +1,21 @@
 import { CanActivate, Injectable, ExecutionContext } from '@nestjs/common';
 import { AuthService } from '../auth.service';
-import { TokenConfig } from '../../utils/constants';
 import { Tokens } from '../../interfaces/token-types';
 import { CookieRequest } from '../../interfaces/cookie-types';
-import { getCookies } from '../../utils/helpers';
+import { getCookies, constants } from '../../utils';
 
 @Injectable()
 export class OnlyAuthenticatedGuard implements CanActivate {
   constructor(private readonly authService: AuthService) {}
 
   public async canActivate(context: ExecutionContext): Promise<boolean> {
+    const { tokenConfig } = constants;
     const request = context.switchToHttp().getRequest<CookieRequest>();
     const cookies = getCookies(request);
     const tokens: Tokens = {
-      refreshToken: cookies[TokenConfig.refreshToken.name],
-      logoutToken: cookies[TokenConfig.logoutToken.name],
-      accessToken: cookies[TokenConfig.accessToken.name],
+      refreshToken: cookies[tokenConfig.refreshToken.name],
+      logoutToken: cookies[tokenConfig.logoutToken.name],
+      accessToken: cookies[tokenConfig.accessToken.name],
     };
     return !!this.authService.validateTokens(tokens);
   }
