+++ packages/api/src/utils/getCookies.ts
@@ -1,6 +1,8 @@
 import { CookieRequest } from '../interfaces/cookie-types';
 
-export const getCookies = (request: CookieRequest) => {
+const getCookies = (request: CookieRequest) => {
   const isDevelopment = process.env.ENV === 'development';
   return isDevelopment ? request.cookies : request.signedCookies;
 };
+
+export default getCookies;
